requirements.txt
alembic==1.12.0
amqp==5.1.1
appdirs==1.4.4
APScheduler==3.10.4
async-timeout==4.0.3
Babel==2.12.1
beautifulsoup4==4.12.2
billiard==4.1.0
blinker==1.6.2
boto3==1.28.43
botocore==1.31.43
Brotli==1.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
commonmark==0.9.1
croniter==1.4.1
crontab==1.0.1
cryptography==41.0.3
decorator==5.1.1
Deprecated==1.2.14
dominate==2.8.0
elastic-transport==8.4.0
elasticsearch==8.9.0
et-xmlfile==1.1.0
firebase-admin==6.4.0
Flask==2.3.3
Flask-Admin==1.6.1
flask-babel==3.1.0
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-Compress==1.13
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-Limiter==3.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
forex-python==1.8
freezegun==1.2.2
frozendict==2.3.8
gekko==1.0.6
google-api-core==2.17.1
google-api-python-client==2.121.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==2.0.2
guess-language-spirit==0.5.3
gunicorn==21.2.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kombu==5.3.2
limits==3.6.0
lxml==4.9.3
mailchimp-transactional==1.0.50
mailchimp3==3.0.21
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
multitasking==0.0.11
numpy==1.25.2
numpy-financial==1.0.0
onetimepass==1.0.1
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.1
pandas==2.1.0
pandas-datareader==0.10.0
passlib==1.7.4
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycparser==2.21
pycryptodome==3.18.0
Pygments==2.16.1
PyJWT==2.8.0
PyQRCode==1.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-slugify==8.0.1
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
redis==5.0.0
requests==2.31.0
rich==13.5.2
rq==1.15.1
rq-scheduler==0.13.1
s3transfer==0.6.2
simplejson==3.19.1
six==1.16.0
soupsieve==2.5
speaklater==1.3
SQLAlchemy==2.0.20
sympy==1.12
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==1.26.16
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.7
wrapt==1.15.0
WTForms==3.0.1
xlrd==2.0.1
yfinance==0.2.3
zipp==3.16.2
